Read correct size for efect description

+++ libeffect.cpp
@@ -66,7 +66,7 @@ int32 scriptlib::effect_set_description(lua_State *L) {
 	check_param_count(L, 2);
 	check_param(L, PARAM_TYPE_EFFECT, 1);
 	effect* peffect = *(effect**)lua_touserdata(L, 1);
-	uint32 v = lua_tonumber(L, 2);
+	uint64 v = lua_tonumber(L, 2);
 	peffect->description = v;
 	return 0;
 }
